@@ -1,30 +1,32 @@
-import React, { Component } from 'react';
-import Button from 'antd/lib/button';
-import Input from 'antd/lib/input';
-import Col from 'antd/lib/col';
-import Modal from 'antd/lib/modal';
-import Router from 'next/router';
-import style from './style.scss';
-import getConfig from 'next/config';
+import React, { Component } from "react";
+import Button from "antd/lib/button";
+import Input from "antd/lib/input";
+import Col from "antd/lib/col";
+import Modal from "antd/lib/modal";
+import Router from "next/router";
+import style from "./style.scss";
+import getConfig from "next/config";
+import { BASE_PATH_URL, isDev } from "../../config";
 const { publicRuntimeConfig } = getConfig();
 
 class HikeSearch extends Component {
   state = {
-    keyword: this.props.keyword ||  '',
+    keyword: this.props.keyword || "",
   };
 
   handleSearchClick = () => {
-    this.handleSubmit(document.getElementById('search').value);
+    this.handleSubmit(document.getElementById("search").value);
   };
 
   handleSubmit = (keyword) => {
     const searchTerm = keyword.trim().replace(/\\|\/|  +/g, " ");
     if (searchTerm.length > 2) {
-      const searchUrl = `/hikes/search/${searchTerm}`;
+      const URL = !isDev ? BASE_PATH_URL : "";
+      const searchUrl = `${URL}/hikes/search?keyword=${searchTerm}`;
       Router.replace(searchUrl);
     } else {
       Modal.error({
-        title: 'Please enter minimum 3 characters',
+        title: "Please enter minimum 3 characters",
       });
       return false;
     }
@@ -38,7 +40,7 @@ class HikeSearch extends Component {
             id="search"
             placeholder="Search"
             className={style.input}
-            onPressEnter={e => this.handleSubmit(e.target.value)}
+            onPressEnter={(e) => this.handleSubmit(e.target.value)}
             autoComplete="off"
             defaultValue={this.state.keyword}
           />
@@ -49,7 +51,10 @@ class HikeSearch extends Component {
             className={style.searchIcon}
             onClick={this.handleSearchClick}
           >
-            <img src={`${publicRuntimeConfig.asset}/static/dist/images/search.svg`} alt="search icon" />
+            <img
+              src={`${publicRuntimeConfig.asset}/static/dist/images/search.svg`}
+              alt="search icon"
+            />
           </Button>
         </Col>
       </div>
